# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SIHIYAN - SIHI Seeds AI Operations Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface StockItem {
  id: number;
  name: string;
  category: string;
  sku: string;
  quantity: number;
  unit: string;
  minQuantity: number;
  /** @nullable */
  maxQuantity?: number | null;
  /** @nullable */
  unitPrice?: number | null;
  /** @nullable */
  supplier?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface StockItemInput {
  name: string;
  category: string;
  sku: string;
  quantity: number;
  unit: string;
  minQuantity: number;
  /** @nullable */
  maxQuantity?: number | null;
  /** @nullable */
  unitPrice?: number | null;
  /** @nullable */
  supplier?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface StockItemUpdate {
  name?: string;
  category?: string;
  sku?: string;
  quantity?: number;
  unit?: string;
  minQuantity?: number;
  /** @nullable */
  maxQuantity?: number | null;
  /** @nullable */
  unitPrice?: number | null;
  /** @nullable */
  supplier?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface StockAdjustInput {
  delta: number;
  reason: string;
}

export interface StockSummary {
  totalItems: number;
  totalValue: number;
  lowStockCount: number;
  outOfStockCount: number;
  categoriesCount: number;
}

export interface CategoryCount {
  category: string;
  count: number;
  totalQuantity: number;
}

export interface Dispatch {
  id: number;
  trackingNumber: string;
  customerName: string;
  /** @nullable */
  customerPhone?: string | null;
  destination: string;
  status: string;
  items: string;
  /** @nullable */
  totalWeight?: number | null;
  /** @nullable */
  totalValue?: number | null;
  /** @nullable */
  dispatchDate?: string | null;
  /** @nullable */
  expectedDelivery?: string | null;
  /** @nullable */
  deliveredAt?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  driver?: string | null;
  /** @nullable */
  vehicle?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface DispatchInput {
  customerName: string;
  /** @nullable */
  customerPhone?: string | null;
  destination: string;
  items: string;
  /** @nullable */
  totalWeight?: number | null;
  /** @nullable */
  totalValue?: number | null;
  /** @nullable */
  dispatchDate?: string | null;
  /** @nullable */
  expectedDelivery?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  driver?: string | null;
  /** @nullable */
  vehicle?: string | null;
}

export interface DispatchUpdate {
  customerName?: string;
  /** @nullable */
  customerPhone?: string | null;
  destination?: string;
  items?: string;
  /** @nullable */
  totalWeight?: number | null;
  /** @nullable */
  totalValue?: number | null;
  /** @nullable */
  dispatchDate?: string | null;
  /** @nullable */
  expectedDelivery?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  driver?: string | null;
  /** @nullable */
  vehicle?: string | null;
}

export interface DispatchStatusUpdate {
  status: string;
  /** @nullable */
  notes?: string | null;
}

export interface DispatchSummary {
  total: number;
  pending: number;
  inTransit: number;
  delivered: number;
  cancelled: number;
  todayDispatches: number;
  totalValueInTransit: number;
}

export interface Warehouse {
  id: number;
  name: string;
  location: string;
  /** @nullable */
  capacity?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface WarehouseInput {
  name: string;
  location: string;
  /** @nullable */
  capacity?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface InventoryItem {
  id: number;
  stockItemId: number;
  warehouseId: number;
  quantity: number;
  /** @nullable */
  batchNumber?: string | null;
  /** @nullable */
  expiryDate?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  stockItemName?: string | null;
  /** @nullable */
  warehouseName?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface InventoryItemInput {
  stockItemId: number;
  warehouseId: number;
  quantity: number;
  /** @nullable */
  batchNumber?: string | null;
  /** @nullable */
  expiryDate?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface InventoryItemUpdate {
  quantity?: number;
  /** @nullable */
  batchNumber?: string | null;
  /** @nullable */
  expiryDate?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface DashboardOverview {
  totalStockItems: number;
  lowStockAlerts: number;
  activeDispatches: number;
  todayDispatches: number;
  totalInventoryValue: number;
  pendingBills: number;
  /** @nullable */
  recentReportDate?: string | null;
}

export interface ActivityItem {
  id: string;
  type: string;
  description: string;
  timestamp: string;
  /** @nullable */
  metadata?: string | null;
}

export interface TrendPoint {
  date: string;
  value: number;
}

export interface TrendData {
  stockTrend: TrendPoint[];
  dispatchTrend: TrendPoint[];
}

export interface DailyReport {
  id: number;
  reportDate: string;
  summary: string;
  /** @nullable */
  stockInsights?: string | null;
  /** @nullable */
  dispatchInsights?: string | null;
  /** @nullable */
  recommendations?: string | null;
  generatedBy: string;
  createdAt: string;
}

export interface OcrScanInput {
  imageBase64: string;
  /** @nullable */
  billType?: string | null;
}

export interface OcrScanResult {
  rawText: string;
  parsedData: string;
  confidence: number;
  /** @nullable */
  billId?: number | null;
}

export interface ScannedBill {
  id: number;
  rawText: string;
  /** @nullable */
  parsedData?: string | null;
  /** @nullable */
  billType?: string | null;
  /** @nullable */
  vendorName?: string | null;
  /** @nullable */
  billAmount?: number | null;
  /** @nullable */
  billDate?: string | null;
  confidence: number;
  createdAt: string;
}

export interface GeminiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface GeminiConversationInput {
  title: string;
}

export interface GeminiMessageInput {
  content: string;
}

export interface GeminiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: GeminiMessage[];
}

export interface GeminiImageInput {
  prompt: string;
}

export interface GeminiImageOutput {
  b64_json: string;
  mimeType: string;
}

export interface GeminiError {
  error: string;
}

export type ListStockItemsParams = {
category?: string;
search?: string;
lowStock?: boolean;
};

export type ListDispatchesParams = {
status?: string;
search?: string;
dateFrom?: string;
dateTo?: string;
};

export type ListInventoryItemsParams = {
warehouseId?: number;
search?: string;
};

export type GetDashboardTrendsParams = {
days?: number;
};