# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SIHIYAN - SIHI Seeds AI Operations Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dispatch:
    get:
      operationId: listDispatches
      tags: [dispatch]
      summary: List all dispatches
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of dispatches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dispatch"
    post:
      operationId: createDispatch
      tags: [dispatch]
      summary: Create a new dispatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispatchInput"
      responses:
        "201":
          description: Created dispatch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispatch"
...
  /dispatch/{id}:
    get:
      operationId: getDispatch
      tags: [dispatch]
      summary: Get a dispatch by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dispatch details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispatch"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateDispatch
      tags: [dispatch]
      summary: Update a dispatch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispatchUpdate"
      responses:
        "200":
          description: Updated dispatch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispatch"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteDispatch
      tags: [dispatch]
      summary: Delete a dispatch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /dispatch/{id}/status:
    patch:
      operationId: updateDispatchStatus
      tags: [dispatch]
      summary: Update dispatch status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispatchStatusUpdate"
      responses:
        "200":
          description: Updated dispatch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispatch"

  # ─── INVENTORY / WAREHOUSE ───────────────────────────────────────────────────
...
  /gemini/conversations:
    get:
      operationId: listGeminiConversations
      tags: [gemini]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeminiConversation"
    post:
      operationId: createGeminiConversation
      tags: [gemini]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeminiConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiConversation"
  /gemini/conversations/{id}:
    get:
      operationId: getGeminiConversation
      tags: [gemini]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
    delete:
      operationId: deleteGeminiConversation
      tags: [gemini]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /gemini/conversations/{id}/messages:
    get:
      operationId: listGeminiMessages
      tags: [gemini]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeminiMessage"
    post:
      operationId: sendGeminiMessage
      tags: [gemini]
      summary: Send a message and receive an AI response (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeminiMessageInput"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}
  /gemini/generate-image:
    post:
      operationId: generateGeminiImage
      tags: [gemini]
      summary: Generate an image from a text prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeminiImageInput"
      responses:
        "200":
          description: Generated image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiImageOutput"
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── STOCK ──────────────────────────────────────────────────────────────────
...
    Dispatch:
      type: object
      properties:
        id:
          type: integer
        trackingNumber:
          type: string
        customerName:
          type: string
        customerPhone:
          type: ["string", "null"]
        destination:
          type: string
        status:
          type: string
        items:
          type: string
        totalWeight:
          type: ["number", "null"]
        totalValue:
          type: ["number", "null"]
        dispatchDate:
          type: ["string", "null"]
          format: date-time
        expectedDelivery:
          type: ["string", "null"]
          format: date-time
        deliveredAt:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        driver:
          type: ["string", "null"]
        vehicle:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - trackingNumber
        - customerName
        - destination
        - status
        - items
        - createdAt
        - updatedAt
    DispatchInput:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: ["string", "null"]
        destination:
          type: string
        items:
          type: string
        totalWeight:
          type: ["number", "null"]
        totalValue:
          type: ["number", "null"]
        dispatchDate:
          type: ["string", "null"]
          format: date-time
        expectedDelivery:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        driver:
          type: ["string", "null"]
        vehicle:
          type: ["string", "null"]
      required:
        - customerName
        - destination
        - items
    DispatchUpdate:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: ["string", "null"]
        destination:
          type: string
        items:
          type: string
        totalWeight:
          type: ["number", "null"]
        totalValue:
          type: ["number", "null"]
        dispatchDate:
          type: ["string", "null"]
          format: date-time
        expectedDelivery:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        driver:
          type: ["string", "null"]
        vehicle:
          type: ["string", "null"]
    DispatchStatusUpdate:
      type: object
      properties:
        status:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - status
...
    GeminiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    GeminiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    GeminiConversationInput:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    GeminiMessageInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    GeminiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/GeminiMessage"
      required:
        - id
        - title
        - createdAt
        - messages
    GeminiImageInput:
      type: object
      properties:
        prompt:
          type: string
      required:
        - prompt
    GeminiImageOutput:
      type: object
      properties:
        b64_json:
          type: string
        mimeType:
          type: string
      required:
        - b64_json
        - mimeType
    GeminiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error